feat(traffic): wireless role by USB + chip-id identity, drop slots/JOIN
Decide the wireless device role at runtime from USB instead of a hardcoded
table, and identify sensors by their own chip id so new boards auto-register
with nothing hardcoded.

- Role by USB host enumeration: master = a PC enumerated the CDC port within
  the settle window, else sensor. Debounced live re-check resets the MCU to
  re-pick (cable pulled -> sensor; sensor plugged into a PC -> master). Drops
  the FICR.DEVICEID -> role table in node_id.c.
- On-air identity = low 32 bits of the chip id (master = reserved id 0); no
  slot, no number, no JOIN handshake. The master auto-registers a sensor in an
  id-keyed registry on first authenticated packet (PSK is the membership gate).
- STATUS uses a synced hashed phase (hash(id, cycle) within the STATUS cycle)
  plus CAD instead of fixed TDMA slots: collision-resistant without
  coordination, re-hashed each cycle so overlaps are transient/self-healing.
- sec_hdr node_id widened 1B -> 4B (nonce/replay updated); PROTO_VER 4 -> 5.
  E/D USB lines report the sensor's 8-hex id; master self-report keeps the "0"
  sentinel and exposes its full 16-hex id on the I line.
- web: parse the I line to show the master's own chip id, fix the diagnostics
  sort for hex ids, sim default to an 8-hex id.
- DESIGN.md updated (sections 2.2/2.3/2.6/2.8/8/9).

Note: PROTO_VER bump means all boards must be reflashed together. Server/DB
unchanged (node_id is already an opaque string). Firmware builds clean in the
CI container (-Wall -Wextra, 0 warnings).

+++ traffic/device/wireless/firmware/src/secure.h
@@ -39,13 +39,13 @@ uint32_t sec_boot_id(void); /* this node's per-boot random id (for diagnostics) 
 /* Seal payload into out[] as [sec_hdr | ciphertext | mac]. node_id is the
  * SENDER's id. Returns the total wire length (SEC_OVERHEAD + payload_len), or
  * <0 if out is too small. Advances this node's tx counter. */
-int sec_seal(uint8_t *out, int out_cap, uint8_t type, uint8_t node_id,
+int sec_seal(uint8_t *out, int out_cap, uint8_t type, uint32_t node_id,
              const void *payload, int payload_len);
 
 /* Parsed cleartext header of a received packet. */
 typedef struct {
     uint8_t  type;
-    uint8_t  node_id;
+    uint32_t node_id;
     uint32_t boot_id;
     uint32_t ctr;
 } sec_meta_t;
